requirements.txt
Django==1.8
Markdown==2.6.1
PyYAML==3.11
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.19
celery==3.1.17
dj-database-url==0.3.0
django-braces==1.4.0
django-extensions==1.5.2
django-filter==0.9.2
django-oauth-toolkit==0.8.0
django-rest-swagger==0.2.9
djangorestframework==3.1.1
factory-boy==2.5.1
fake-factory==0.5.0
gunicorn==19.3.0
kombu==3.0.24
mock==1.0.1
oauthlib==0.7.2
psycopg2==2.6
pytz==2015.2
six==1.9.0
sure==1.2.10
whitenoise==1.0.6
